+++ common/src/main/java/org/astraea/common/metrics/MBeanRegister.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.common.metrics.jmx;
+package org.astraea.common.metrics;
 
 import java.lang.management.ManagementFactory;
 import java.util.ArrayList;
